pyproject.toml
[project]
name = "regulation-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "bs4>=0.0.2",
    "datasets>=3.6.0",
    "dotenv>=0.9.9",
    "huggingface-hub>=0.33.0",
    "ipykernel>=6.29.5",
    "jpype1>=1.5.2",
    "langchain>=0.3.25",
    "langchain-chroma>=0.2.4",
    "langchain-community>=0.3.25",
    "langchain-core>=0.3.65",
    "langchain-unstructured>=0.1.6",
    "lxml>=5.4.0",
    "openai>=1.86.0",
    "openpyxl>=3.1.5",
    "pdf2image>=1.17.0",
    "pdfplumber>=0.11.7",
    "pi-heif>=0.22.0",
    "pillow>=11.2.1",
    "pymupdf>=1.26.1",
    "ragas>=0.2.15",
    "sentence-transformers>=4.1.0",
    "unstructured[pdf]>=0.17.2",
    "unstructured-inference>=1.0.5",
    "unstructured-pytesseract>=0.3.15",
    "python-pptx>=1.0.2",
    "docx2txt>=0.9",
    "python-docx>=1.1.2",
]
